+++ src/extensions/index.js
@@ -39,6 +39,7 @@ import CodeBlock from './code-block'
 import Columns from './columns'
 import Datetime from './datetime'
 import { CrossReference, DocumentReferences } from './document-references'
+import { Pagination } from './pagination'
 import Echarts from './echarts'
 import File from './file'
 import FileHandler from './file-handler'
@@ -163,6 +164,7 @@ export const getDefaultExtensions = ({ container, options, uploadFileMap }) => {
     bookmark: Bookmark,
     crossReference: CrossReference,
     documentReferences: DocumentReferences,
+    pagination: Pagination,
     footnote: [Footnotes, FootnoteReference, Footnote],
     'hard-break': BreakMarks.configure({
       visible: page?.showBreakMarks,